+++ src/main/java/com/bitala/api/mantenimiento/controllers/UnidadController.java
@@ -5,15 +5,9 @@ import org.springframework.web.bind.annotation.*;
 import com.bitala.api.mantenimiento.models.Unidad;
 import com.bitala.api.mantenimiento.repository.IUnidadRepository;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import java.net.URI;
 import java.util.List;
-import java.util.Optional;
 
 /**
  * API MANTENIMIENTOS - BITALA
@@ -33,9 +27,6 @@ public class UnidadController {
     @Autowired
     private IUnidadRepository unidadRepository;
 
-    //Objeto Optional<Unidad>
-    Optional<Unidad> unidadOptional;
-
     //Constructor para inyección de dependencias
     public UnidadController(IUnidadRepository unidadRepository) {
         this.unidadRepository = unidadRepository;
@@ -49,56 +40,31 @@ public class UnidadController {
 
     //Busca una Unidad por id
     @GetMapping("/{id}")
-    public ResponseEntity<Unidad> findById(@PathVariable("id") Long id){
-        unidadOptional = unidadRepository.findById(id);
-        return unidadOptional.map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
+    public Unidad findById(@PathVariable("id") Long id){
+       if(unidadRepository.existsById(id)) return unidadRepository.findById(id).orElse(null);
+       else return null;
     }
 
     //Agrega una nueva Unidad
     @PostMapping
-    public ResponseEntity<Unidad> createUnidad(@RequestBody Unidad unidad){
-        try {
-            Unidad nuevaUnidad = unidadRepository.save(unidad);
-            return ResponseEntity.created(URI.create("/api/unidad/" + nuevaUnidad.getIdUnidad())).body(nuevaUnidad);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
+    public Unidad createUnidad(@RequestBody Unidad unidad){
+        return unidadRepository.save(unidad);
     }
 
     //Modifica una Unidad por id
     @PutMapping("/{id}")
-    public ResponseEntity<Unidad> updateUnidad(@PathVariable Long id, @RequestBody Unidad unidadData){
-        try {
-            unidadOptional = unidadRepository.findById(id);
-            if (unidadOptional.isPresent()) {
-                Unidad unidadExistente = unidadOptional.get();
-                BeanUtils.copyProperties(unidadData, unidadExistente, "id");
-                Unidad unidadActualizada = unidadRepository.save(unidadExistente);
-                return ResponseEntity.ok(unidadActualizada);
-            } else {
-                return ResponseEntity.notFound().build();
-            }
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
+    public Unidad updateUnidad(@PathVariable Long id, @RequestBody Unidad unidadData){
+        Unidad unidad = unidadRepository.findById(id).orElse(null);
+
+        if(unidad != null){
+            unidad.setIdEmpresa(unidadData.getIdEmpresa());
+            return unidadRepository.save(unidad);
+        }else return null;
     }
 
     //Elimina una Unidad por id
     @DeleteMapping("/{id}")
-    public ResponseEntity<Unidad> deleteUnidad(@PathVariable("id") Long id){
-        try {
-            if(unidadRepository.existsById(id)) {
-                unidadRepository.deleteById(id);
-                return ResponseEntity.noContent().build();
-            } else return ResponseEntity.notFound().build();
-        } catch (Exception e){
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-    }
-
-    // Manejo de excepciones genéricas para cualquier otra excepción no capturada
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleGenericException(Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+    public void deleteUnidad(@PathVariable("id") Long id){
+        if(unidadRepository.existsById(id)) unidadRepository.deleteById(id);
     }
 }
